feat: optimize parsing invalid chars
Image refs with many and long subdomains that have invalid characters
in the image reference cause the regex engine to get bogged down
and could become a regular expression denial of service attack.

For example `docker.artifactory.us.foo.mycompany.com/bar/node?18` takes >4mins to parse and return as an invalid reference. With the updates in this change set the whole test set runs in 2ms.